pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skillkit"
version = "0.1.0"
description = "A Streamlit dashboard for managing AI agent skill files"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.28.0",
    "PyYAML>=6.0",
    "pandas>=2.0",
    "markdown>=3.5",
]

[project.optional-dependencies]
analysis = ["sentence-transformers>=2.2.0", "numpy>=1.24.0"]
dev = ["pytest>=7.0"]

[project.scripts]
skillkit = "skillkit.app:_cli"

[tool.setuptools.packages.find]
include = ["skillkit*"]
